requirements/dev.txt
# This file pulls in everything a developer needs. If it's a basic package
# needed to run the site, it belongs in requirements/prod.txt. If it's a
# package for developers (testing, docs, etc.), it goes in this file.

django-debug-toolbar
django-extensions

-r prod.txt
-r ../vendor/src/funfactory/funfactory/requirements/compiled.txt
